requirements.txt
streamlit
pandas
beautifulsoup4
requests
python-docx
fpdf
